Move git configuration to conventional location
I used to keep git config at .config/git/config (according to XDG
recommendation, which says that all configuration should by default
live in ~/.config).  I still think it's a good thing to do (as it
helps to keep your HOME less cluttered); however, almost everyone
will be looking for git configuration in ~/.gitconfig - and it would
be contrary to the purpose of this repo to confuse them.