Git config: remove superfluous aliases
I personally prefer aliasing commit to 'k' and checkout to 'c', but it
seems that the rest of the world consistently uses 'ci' and 'co'
aliases, so it would be contrary to the purpose of this repo to force
non-standard config.  Similarly with 'branch' aliases.

Also, remove one grep alias and simple log alias (as pretty logs defined
below are much much better anyway).